requirements.txt
absl_py==1.4.0
einops==0.6.1
matplotlib==3.7.1
numpy==1.24.2
scipy==1.11.2
torch==1.13.1
tqdm==4.65.0
wandb==0.13.10
